Optimised heroku-local networking, added Postgres
Did you use Linter?
Yes

1. What exactly did you do, explain step by step using an ordered list?

a. Changed the port forwarding from guest to host to 5000 instead of 8000 as the former is the default for heroku local web command.
b. Added the installation of postgresql and its necessary dependencies in the vagrant provisioning.

2. Why did you do it? Any comments?

Now you can run heroku local web and access the local server running at localhost:5000 in your browser. This separates local testing and deployment but allows you to launch
local testing directly from within the vagrant ssh and inside the virtualenv.